requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Security-Too==5.5.2
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0
google-api-python-client==2.149.0
pytesseract==0.3.13
Pillow==11.0.0
psycopg2-binary==2.9.10
gunicorn==21.2.0